refactor(fjagepy): fold message registry into Message
Move the registry and register_message back into Message.py, where the
Message subclass check needs no import gymnastics, and drop the
messageregistry module. Gateway.registerMessage keeps the fully qualified
name check the gateway spec requires; register_message accepts any
non-empty name so @message and MessageClass can pass through.

Bare @message now registers under the plain Python class name again
rather than module.Class, which put __main__ paths on the wire.

Also drop the unused Gateway._MESSAGE_REGISTRY attribute.
